requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appdirs==1.4.4
asn1crypto==1.5.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
Bottleneck @ file:///Users/runner/miniforge3/conda-bld/bottleneck_1656803827350/work
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clikit==0.6.2
cmdstanpy @ file:///home/conda/feedstock_root/build_artifacts/cmdstanpy_1661515366871/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
commonmark==0.9.1
convertdate @ file:///home/conda/feedstock_root/build_artifacts/convertdate_1642883757836/work
crashtest==0.3.1
cryptography==36.0.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython @ file:///Users/runner/miniforge3/conda-bld/cython_1659102017892/work
decorator==5.1.1
entrypoints==0.4
ephem @ file:///Users/runner/miniforge3/conda-bld/ephem_1652789194763/work
filelock==3.8.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1661381414837/work
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
hijri-converter @ file:///home/conda/feedstock_root/build_artifacts/hijri-converter_1653348285237/work
holidays @ file:///home/conda/feedstock_root/build_artifacts/holidays_1661252738306/work
httpstan==4.6.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.14.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1657953277494/work
korean-lunar-calendar @ file:///home/conda/feedstock_root/build_artifacts/korean_lunar_calendar_1589354365811/work
LunarCalendar==0.0.9
MarkupSafe==2.1.1
marshmallow==3.17.1
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1661440102531/work
multidict==6.0.2
munkres==1.1.4
numexpr @ file:///Users/runner/miniforge3/conda-bld/numexpr_1658076484743/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1649806371302/work
oscrypto==1.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.3.5
pastel==0.2.1
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1660385974720/work
pkgutil_resolve_name==1.3.10
prophet @ file:///Users/runner/miniforge3/conda-bld/prophet_1661461864011/work
protobuf==3.20.1
pyarrow==8.0.0
pycparser==2.21
pycryptodomex==3.15.0
pydeck==0.8.0b1
Pygments==2.13.0
PyJWT==2.4.0
pylev==1.4.0
PyMeeus==0.5.11
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyrsistent==0.18.1
pysimdjson==3.2.0
pystan==3.3.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.19.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1660580468234/work
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.5.1
semver==2.13.0
setuptools-git==1.2
simplejson==3.17.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
snowflake-connector-python==2.7.12
streamlit==1.12.2
streamlit-aggrid==0.2.3.post2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
tzdata==2022.2
tzlocal==4.2
ujson @ file:///Users/runner/miniforge3/conda-bld/ujson_1653057579360/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1649112008529/work
urllib3==1.26.12
validators==0.20.0
webargs==8.2.0
yarl==1.8.1
zipp==3.8.1
